Screwy restore process for snapshots
Created because we don't have real backups.  Just snapshots.  Using docs with VG and Disk info for all of our machines that is updated daily, cross references disk and vg info with output from 3PAR SANs to determine what server/volume group a presented snapshot is from.  This is done using only immutable data because our storage team cannot be trusted to use a naming convention or even name the snapshot at all. (seriously, they have presented us volumes with no names before)  It recreates the VG, mounts all the file systems, creates an nfs export, mounts it on the appropriate server in /RESTORE, and creates an atjob that will undo this entire process and clean everything up in 24 hours.